Improvements to edit columns dialog.
Changes are applied instantly and it can respond to the columns changing mid-edit.
It also maintains its state from one opening to the next (sorting, filter string)

This seemed like the best way to fix a bug (things would break if the columns
changed whilst you were editing which were visible. This was because the dialog
exec_() method doesn't actually bloick the main event loop, in a sense it is just
a copy of that event loop. So I couldn't easily stop new shots arriving and
messing with the column structure whilst we were editing it).

Just making the whole thing dynamic seemed like a good fix and is nice
behaviour too.

Added context menu to show and hide selection in the dialog box too.