Fixes #16, routines not reordering correctly with GUI buttons.
Problem was confusing a passive transformation with an active transformation:

passive transformation: constructing a new list from a list of indices saying
where in the old list each item should come from (what we were doing)

active transformation: constructing a new list out of a list of indices for
where in the new list items in the old list should be moved to (what we should
have been doing)

When a single routine was moved, this simply swapped it with the one above or
below it and so a passive and active transformation of the order are identical.
But for reordering multiple routines at once it gave the wrong result.